labs/devnet-workshop/rest-api-workshop/webex/webhook.py
from flask import Flask, request
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

@app.route("/webhook", methods=["POST"])
def webhook():
    body = request.json
    print(body)
    return "OK"
