labs/devnet-workshop/rest-api-workshop/webex/requirements.txt
backoff==2.2.1
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
coloredlogs==15.0.1
flask==3.0.3
future==1.0.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==2.1.5
pyjwt==2.10.1
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==2.4.0
webex-bot==0.5.2
webexpythonsdk==2.0.4
webexteamssdk==1.6.1
websockets==11.0.3
werkzeug==3.0.3
